Support large Telegram user IDs in base schema

+++ database/schema.sql
@@ -5,7 +5,7 @@ CREATE EXTENSION IF NOT EXISTS pgcrypto;
 
 CREATE TABLE users (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    telegram_user_id BIGINT NOT NULL UNIQUE,
+    telegram_user_id NUMERIC(20,0) NOT NULL UNIQUE,
     display_name TEXT NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
